move Encryption logic from MatrixRTCSession into EncryptionManager
